tui/pyproject.toml
[project]
name = "memex-tui"
version = "0.1.0"
description = "Interactive TUI control plane for memex and dagit"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.50",
    "openai>=1.0",
    "httpx>=0.25",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dagit = [
    "dagit @ git+https://github.com/systemshift/dagit.git",
]

[project.scripts]
memex-tui = "memex_tui:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
